Switch internal storage from list to dict
Previously, when phones_for_word() was called, we walked the entire
dictionary looking for matching words. This was, to put it mildly, slow.
This commit changes the internals to use a dictionary instead, which
speeds this up from ~30ms to less than a microsecond on my machine.

There is a concomitant slowdown in search() and related functions of
around 30-40%.

We also have to change one of the tests which relied on the old internal
representation.